+++ modules/findimage.py
