+++ modules/interwikilist/__init__.py
@@ -5,13 +5,13 @@ from os.path import abspath
 
 def iwlist():
     cp = ConfigParser()
-    cp.read(abspath("./plugins/interwikilist/list.cfg"))
+    cp.read(abspath("./modules/interwikilist/list.cfg"))
     section = cp.sections()[0]
     return (cp.options(section))
 
 
 def iwlink(iw):
     cp = ConfigParser()
-    cp.read(abspath("./plugins/interwikilist/list.cfg"))
+    cp.read(abspath("./modules/interwikilist/list.cfg"))
     section = cp.sections()[0]
     return (cp.get(section, iw))
