+++ modules/rc/__init__.py
@@ -2,8 +2,8 @@
 import aiohttp
 import json
 
-from plugins.UTC8 import UTC8
-from plugins.pbc import pbc2
+from modules.UTC8 import UTC8
+from modules.pbc import pbc2
 
 
 async def rc():
